Fix a bug that caused buttons in "Languages and Input" dialog to be truncated.
Change the layout so the input method configure button will be shown right
next to the input method name (the layout was bad as the row was sharing the
same columnset with the huge the "Display Google Chrome in this language"
button).

Also widen the width of the dialog from 96 to 106, so the "Display Google
Chrome in this language" button does not get truncated. To do this, define
separate dialog size settings for this dialog.

BUG=chromium-os:3235
TEST=manually

Review URL: http://codereview.chromium.org/2106008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@47404 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: aeeb9f1804ece6b506a747d939d40be5817420ce