Removed the "Reload" button from the HTTP cache tab of the net-internals page.
The "Reload" button was used to refresh the stats + cache listings.
With this change, I poll the statistics for updates, so the reload button is unecessary.

Note that the cache listing is too expensive to poll, so I replaced it with a link to a listings page.

Lastly, updated the text dump to include the http cache statistics.

BUG=37421

Review URL: http://codereview.chromium.org/2085005

git-svn-id: http://src.chromium.org/svn/trunk/src@47499 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: bc2c1e7f89027ca1036c83491c71990eed4bc681