Pull latest ppapi for initial testing support, implement checks so the tests pass in Chrome. Mostly, this involves better error checking.
The most substantial change is that I made TransportDIB able to return NULL on failure to make a PlatformCanvas, which we run into when makeing too large of a canvas from a plugin. I checked all the callers of this function to make sure they all handled the problem (many already did).

TEST=pulls Pepper test
BUG=-none
Review URL: http://codereview.chromium.org/2101004

git-svn-id: http://src.chromium.org/svn/trunk/src@47526 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 745e6e901d87efe02ca1b262fff11f41efb7ba5c