Switch UITestBase::WaitForFinish to use WaitUntilCookieNotEmpty
The way plugin tests (at least) use the completion cookie is to set the status
to either the 'OK' value, or to an error string, then signal completion. With
the current impletation of WaitForFinish (which is to wait until the cookie is
set to 'OK'), the tests time out with no information. If we instead wait for any
value and then assert that it's the expected value, failures will:
a) happen immediately instead of causing a timeout (so cycle times will be faster), and
b) provide useful failure messages for debugging the failure.

BUG=none
TEST=UI tests still work correctly; failures log useful messages.

Review URL: http://codereview.chromium.org/2119003

git-svn-id: http://src.chromium.org/svn/trunk/src@47566 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 99ef3b90b6eaddcc2561b9bf9f5171781535db6a