Implement fading in for the InfoBubble base class.
For bubbles that are not shown as a direct result of
a mouse click, we use fade in. This applies to
the FirstRun bubble and the extension installed
bubble.

For bubbles that have controls (buttons/links) we
fade out if interacting with those controls closes
the bubble. This applies to the first run bubble,
app launcher, extension installed bubble, bookmark
bubble and the content settings bubble.

BUG=None
TEST=Make sure the above-mentioned bubbles fade in
and out during hide and show as described above.

Review URL: http://codereview.chromium.org/2079008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@47575 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4d4fdaf5bcdc45b777a29606803294a54a7c27e3