Implement fading in for the InfoBubble base class.
For bubbles that are not shown as a direct result of
a mouse click, we use fade in. This applies to
the FirstRun bubble and the extension installed
bubble.

For bubbles that have controls (buttons/links) we
fade out if interacting with those controls closes
the bubble. This applies to the first run bubble,
app launcher, extension installed bubble, bookmark
bubble and the content settings bubble.

BUG=None
TEST=Make sure the above-mentioned bubbles fade in
and out during hide and show as described above.

Review URL: http://codereview.chromium.org/2079008

git-svn-id: http://src.chromium.org/svn/trunk/src@47575 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f7177a7020200aac477c63011c8864a1da62262a