Fix the ChromeFrame new window popup test by ensuring that CreateExternalTabComplete executes on the IPC thread as before and InitializeComplete executes as a task on the UI thread. The test failed because we receive IPCs for the new tab before we have a tab proxy for it.
To fix the crash in bug http://code.google.com/p/chromium/issues/detail?id=44245, we set the automation_server_ member to NULL
after ReleaseAutomationServer returns. There are a number of races related to the automation server being released which
should hopefully go away if we remove dependency on the AutomationProxy object. Added a TODO to that effect in the code.

Bug=44245

Review URL: http://codereview.chromium.org/2079012

git-svn-id: http://src.chromium.org/svn/trunk/src@47612 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d934731686ee8df3c5d859f9e9546eda7db970de