fix chrome.tabs.onUpdated bugs, add browsertest.
This CL addresses a few issues related to the behavior of the onUpdated event.

Issue 1: The page re-entering the load state and the completing was causing multiple 'complete' status events to be fired. We now only report the first 'complete' after the didNavigate message is fired (iframe navigation, for example).

Issue 2: We were initializing the URL when the TabEntry was created, and this caused us to fail to send the url with the first navigation because we thought it wasn't changing.

BUG=27208,37149

Review URL: http://codereview.chromium.org/2111010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@47717 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c7f6ae184256336e4d26912076b550a22e93b474