fix chrome.tabs.onUpdated bugs, add browsertest.
This CL addresses a few issues related to the behavior of the onUpdated event.

Issue 1: The page re-entering the load state and the completing was causing multiple 'complete' status events to be fired. We now only report the first 'complete' after the didNavigate message is fired (iframe navigation, for example).

Issue 2: We were initializing the URL when the TabEntry was created, and this caused us to fail to send the url with the first navigation because we thought it wasn't changing.

BUG=27208,37149

Review URL: http://codereview.chromium.org/2111010

git-svn-id: http://src.chromium.org/svn/trunk/src@47717 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b5ea8d0c6a5d0868a7ee8e9e90a4edefaaa94232