chromeos: Add _CHROME_LOGGED_IN property on root window.
This gets set to either 1 or 0 depending on whether the user
is currently logged in or not.  I'm doing this as a step towards
using the same window manager process during and after login;
this will let the WM know which log file it should be using, for
instance.

BUG=chromium-os:2790
TEST=tried it, and will add an autotest for this soon

Review URL: http://codereview.chromium.org/2078017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@47748 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 61b4528ed73988c77f60d294f7381f7c98c33fe4