On ARM, only blit the needed rects from the TransportDIB.
See the comments in the bug: this is an important code path for
performance and we've found that different code works best for
notebook platforms, but regresses on desktops/laptops.

We could implement a lot of complex code to measure the speed at
run-time and select the best algorithm, but it appears that we can get
good results from just an #ifdef, so we'll start simple.

BUG=44124

http://codereview.chromium.org/2110022/show

git-svn-id: http://src.chromium.org/svn/trunk/src@48058 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5b525a472c342dc4ce97c10695c542184c543b10