AutoFill sending forms should include elements with autoComplete="off" attributes.
On these pages we were hitting the DCHECK_EQ in autofill_manager.cc:237 when filling these forms.  This was due to the lack of <input> elements with
autoComplete="off" attributes set when originally sending the form data with the RenderView::SendForms() call.  This changes the renderer to add
these elements when originally sent and processed by the AutoFillManager::FormsSeen() call.

BUG=40700, 41702
TEST=Manual tests of pages in bug reports.

Review URL: http://codereview.chromium.org/2135012

git-svn-id: http://src.chromium.org/svn/trunk/src@48174 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0476e36a253f9de6580e4c4fb7d5a15985dac6fc