GetEnabledExtensions should also return Loaded Extensions.
We use to not need them since we would simply overinstall them all the time.
But now that overinstall is allowed and reset the extension state (e.g., close and reopen the background page), we can't simply install exploded extensions at will anymore, we must first check if it was already installed or not (which means GetEnabledExtensions must return them) and only install them when they are not already there.

BUG=0
TEST=none

Review URL: http://codereview.chromium.org/2303004

git-svn-id: http://src.chromium.org/svn/trunk/src@48400 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7a8b5d6fb60fde0dcc3b0483fce8a94b0fd065f2