Adding a TearDown() method to class LiveSyncTest.
The LiveSyncTest class derives from InProcessBrowserTest and overrides
the SetUp method to account for the case when a local sync server needs
to be started. However, there isn't a corresponding override for the
TearDown method that stops the local sync server. This changelist adds
an overload for TearDown that stops the local sync server if one was
started in SetUp.

BUG=45181
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/2260001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@48457 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1e8ea276897199ba1967a5ac678f73a0a371aceb