First pass at experimental omnibox API. There are plenty of rough edges and features to work on, but it's in a usable state.
When an extension is installed that specifies an omnibox keyword in its manifest, we add that keyword to the user's list of Search Engines. The user can then edit this keyword later.

I'm leveraging most of the original search engine keyword code. An extension keyword has a special URL that identifies it as an extension keyword. There is some special case code to treat these keywords slightly differently throughout
the omnibox code.

BUG=38884

Review URL: http://codereview.chromium.org/2078021

git-svn-id: http://src.chromium.org/svn/trunk/src@48503 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 441cd668de630316a99d09ec58634edacb575088