GTK: Fix padding of the page/app menus.
In r47478, I removed the hbox that the page and app menus were in because I
though that with the new left/right toolbar boxes, they were unnecessary.
But they're needed to group the two buttons together with no space between
them, but with padding on the right.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/2375002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@48522 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1b86afce7b4033c5a48afe3025b57e123b409566