Fix extension context menu API onclick handlers.
I believe the following CL had accidentally broken them:

http://src.chromium.org/viewvc/chrome?view=rev&revision=47972

I haven't yet added end-to-end tests for the context menu API that would catch
problems like this, but I just created a bug for that (crbug.com/45811) and 
made the bug for bringing the API out of experimental blocked on it.

I also fixed a small >80 chars style nit in this CL.

BUG=45741
TEST=The onclick handler for an extension-added context menu item should now
be called. See the bug for a minimal code snippet to put into a background page
for this.



Review URL: http://codereview.chromium.org/2626001

git-svn-id: http://src.chromium.org/svn/trunk/src@48936 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b713abf746d56644675e0ba64f7a7ca55d070fa3