Fix audio "clicking" for strongbad demo
The strongbad demo plays an audio file of 11kHz. However we hardcoded
in the AudioRendererHost to always open an audio device of 8k samples
which happens to be too big for streams of 11khz. This makes the IPC
transport packet size much smaller than the hardware buffer size and
hardware packet is always partially filled.

This change allow the AudioRendererHost to designate the transport
packet size based on the hardware packet size. The hardware packet
size is now adjusted based on the sample rate.

BUG=46007
TEST=http://smokescreen.us/demos/sb45demo.html
The above audio plays fine.

Review URL: http://codereview.chromium.org/2651001

git-svn-id: http://src.chromium.org/svn/trunk/src@49078 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 11c5cf0323d6e06ef1cdec215e601f4b783f2e78