Don't save last used locations in open and save dialogs when in incognito mode.
For "Save page as..." we simply don't store the save location in incognito mode.

For "Open File"... and file upload controls, it's a bit more complicated: because NSOpenPanel and NSSavePanel default to using the last chosen directory from NavigationServices (which we can't influence),
we provide a preference value prefs::kLastChosenDirectory as fallback there, which is not persisted in incognito mode.

BUG=45064
TEST=See bug for manual test.

Review URL: http://codereview.chromium.org/2450004

git-svn-id: http://src.chromium.org/svn/trunk/src@49237 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e49615d735a3c991a4e79c27e06848e9662aece6