AutoFill Profiles dialog implemented according to new mocks on Mac
New mocks are attached to bug 44622.  These changes replace the in-place editing of address and credit card records with a table of records and separate sheets for manipulating the record data.  Changes to the layout of fields on the sheets has been done also.
AutoFillDialog.xib changes:  Replaced disclosure based list of address and credit cards with an NSTableView of the same data.  Added buttons for "Add", "Edit", and "Remove".
Replaced AutoFillAddressViewController.xib with sheet-based AutoFillAddressSheetController.xib.
Replaced AutoFillCreditCardViewController.xib with sheet-based AutoFillCreditCardSheetController.xib.

BUG=44621
TEST=AutoFillAddressModelTest,AutoFillAddressSheetControllerTest,AutoFillCreditCardModelTest,AutoFillCreditCardSheetControllerTest,AutoFillDialogControllerTest

Review URL: http://codereview.chromium.org/2673006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@49274 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6de9bceee1c01f60ff0679afc84ff7126a4ab41e