NaCl: Remove call to NaClAppDtor()
Calling NaClAppDtor() is unnecessary because the process is about to
exit anyway, and it happens only in the error path because
NaClAppDtor() is not able to shut down running NaCl threads safely.
There is a TODO for its removal in the NaCl tree.

This is related to http://codereview.chromium.org/2659002.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=560
TEST=nacl_ui_test

Review URL: http://codereview.chromium.org/2753002

git-svn-id: http://src.chromium.org/svn/trunk/src@49324 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: fc91e2175544275c1e382138470abfc4feeb7011