Silently discard invalid OnFollowDeferredRedirect requests sent by the renderer to the browser.
It is a bug if the renderer sends these, but this check prevents the browser from crashing as a consequence of it.
One situation where this can happen today (and needs to be fixed renderer-side), is when the renderer cancel's a request in response to IGNORE policy, but then later tries to follow the redirect on the cancelled request.

BUG=20470


Review URL: http://codereview.chromium.org/2793001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@49421 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1fa3017ced7abc8bf9f30389eba438c78de79c07