Fix and re-enable RedirectTest.ClientCancelled.
The hack to create and dispatch an event after a timeout to fool the browser
into thinking it is user initiated stopped working after WebKit r49827.

Changing the test to navigate to a javascript: url instead, which qualifies
as a user initiated location change.

BUG=45053
TEST=ui_tests --gtest_filter=RedirectTest.ClientCancelled

Review URL: http://codereview.chromium.org/2225007

git-svn-id: http://src.chromium.org/svn/trunk/src@49681 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 62ef3ba442b9c9b9f0d7976ad6e10080e9843908