Fix a regression that accesses file system synchronously in URLRequestFileJob::Factory.
This regression came from revision 5944. In addition to
performance penalty, the change in 5944 introduced a new
issue: the FileJob -> FileDirJob redirection code was
not hit for directory path without trailing slash. Revision
5999 tried to fix that with a questionable approach. 

The meat of this change is reverting the relevant code in
5944 and 5999. Due to the reverting, this change also
need to have the following code changes:

(1) revision 10408 fixes a bug on "file:///". We don't
regress the fix but the scenario goes through a different
code part. Some comments are changed accordingly and a
unit test is added.

(2) revision 23944 fixes a related crash. The code change
in 23944 is still valid but the unit test needs to be
rewritten.


BUG=43116,1474,18686
TEST=URLRequestTest.FileDir*

Review URL: http://codereview.chromium.org/2748004

git-svn-id: http://src.chromium.org/svn/trunk/src@49737 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a9cd9931ac915e1b560ba1bdf836421ac5fb038a