Mac/clang: Possibly contentious changes.
As discussed on irc, we're getting rid of const NSObjects, and we're making all properties nonatomic.

const: All of cocoa takes nonconst NSObjects, and passing a const NSString to a function that takes NSString is a const violation. gcc doesn't complain about this, but clang intentionally does. constness is usually done via immutable base classes and mutable subclasses in cocoa anyway, so getting rid of const isn't that bad. Also, we don't really have a choice.

nonatomic: Some of our classes have custom setters that are not @synchronized. If the @property for that is not non-atomic, then the interface claims that the method is synchronized but the implementation actually isn't. That's a bug. gcc happens not to complain about this, but clang does. Since we shouldn't need atomic properties anywhere, the simple rule is now to just make all properties nonatomic.

Review URL: http://codereview.chromium.org/2769014

git-svn-id: http://src.chromium.org/svn/trunk/src@49808 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7a6081326ce4e6c122e94201e9bbef8ed86c1f40