Nib change: chevron icon for "off the side" button on bookmark bar now bottom-aligned.
Technically this does not center the icon like the bug asks; it
bottom-aligns it.  However, the image in the icon itself is
top-aligned (has whitespace on the bottom).  This change makes the
chevron line up with the center of bookmark button text next to it so
it looks correct.

BUG=http://crbug.com/45740

Review URL: http://codereview.chromium.org/2870003

git-svn-id: http://src.chromium.org/svn/trunk/src@49874 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 742cb7e56ec060281079aa0c3bd659265c0b1eaa