Nib change: chevron icon for "off the side" button on bookmark bar now bottom-aligned.
Technically this does not center the icon like the bug asks; it
bottom-aligns it.  However, the image in the icon itself is
top-aligned (has whitespace on the bottom).  This change makes the
chevron line up with the center of bookmark button text next to it so
it looks correct.

BUG=http://crbug.com/45740

Review URL: http://codereview.chromium.org/2870003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@49874 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d5662f3d4a71aca7d7fd16a357bda7d998337597