Allow randomized search engines in selection dialog.
Add an option in the master_preferences to randomize the order of logos in the search engine selection screen.  This allows us to cull out a small subset of users to take part in a test of the effects of logo position on search engine choice.

I also filed http://crbug.com/46606 to remind me to revert this change when testing is complete.

BUG=46302
TEST=Add the following options to master_preferences:

 "distribution": {
    "search_engine_experiment": true,
    "search_engine_experiment_randomize": true
  }

Run a few times as --first-run, and see search engines appear in random order in the selection box.

Review URL: http://codereview.chromium.org/2808006

git-svn-id: http://src.chromium.org/svn/trunk/src@49950 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 448fbeaeb326d23988e1e7efd16cb74aa6d46c88