Map Liberation Sans Mono and Ascender Sans Mono to Courier New. They're metrically compatible with each other, but the former is sans serif (monospace) while ther latter is serif (monospace), which is why we didn't include this mapping in the two previous CLs (one by Evan and the other by me).
However, on ChromeOS, we use 'Ascender Sans Mono' as a substitute for Courier
New so that they have to be regarded as 'identical'. 

Without this change, a text node  styled with 'Courier New' is not rendered
with our replacement font while a text node styled with 'Courier New, momospace'
or 'monospace' is rendered with our replacement font. 

BUG=cros:4006
TEST=See the bug.

Review URL: http://codereview.chromium.org/2838003

git-svn-id: http://src.chromium.org/svn/trunk/src@49975 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0c11b9da3f19686d1ac3a4e1fb5c7fc7e96fde20