Landing CL http://codereview.chromium.org/2767013/show on behalf of petkov@chromium.org:
Add one more Chrome OS hardware class MetricsLog unit tests. 

Since the hardware class attribute is added to the XML when the log is 
closed, it's important to test for its presence in the encoded string 
when the log contains some events. The existing unit test covers only 
the case when the log is empty. 

BUG=none 
TEST=ran unit tests 


Review URL: http://codereview.chromium.org/2814005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@49980 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 77e441fa08e8590395c769025874b2b2c9062b2c