This CL fixes issue 43800 - RTL/LTR eliding in Linux omnibox dropdown.
Unlike Windows, in Linux, we do not draw the strings in autocomplete popup by ourselves. Instead, we rely on Pango to draw the strings. We only need to pass the text, the attribute, and the width to Pango, and Pango will draw the text correctly (taking care of visual runs for bidi text).

The only small problems are that:
when the UI is RTL, for pure English text, we need to insert LRE mark so that the ending punctuation is drawn correctly as is (not flipped), for example "....(choose location)" will not be drawn as "(....(choose location" in RTL UI.

Also, for some (not all) pure English text, when eliding, the ellipsis appears at the very left of the string in RTL UI. I do not know the reason of such inconsistency, but looks like adding LRE mark to pure English also forces the ellipsis appears at the right of the string.


BUG=43800
TEST=Omnibox should elide long URLs, mixed LTR/RTL text, etc. correctly in both LTR and RTL locales
Review URL: http://codereview.chromium.org/2824027

git-svn-id: http://src.chromium.org/svn/trunk/src@50773 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c1991789112d8772916f8db5b620d2639e8bbd7e