Do some minor cleanup in histogram.*
I was trying to debug some code, and noticed some excess
complexity.

There are no significant semantic changes.
DCHECK on a foo.get() is a waste given we're about to deref
the pointer.
Execution cost should be be minially less, and we get rid
of several lines of code.

BUG=46263
r=mbelshe
Review URL: http://codereview.chromium.org/2805033

git-svn-id: http://src.chromium.org/svn/trunk/src@50847 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 8b0c46a1dd71082b48c5e42ec93c92dd041309f2