Revert 50913 TranslateManagerTest.NormalTranslate fails on Linux
TranslateInfoBarDelegate2 were leaked. (Yikes!)
This revealed that the NormalTranslate unit-test was incorrect (it was not testing that changing the source or target language from the after translate infobar causes a translation).
This also revealed that returning reference string in the delegate is not a brilliant idea as when we switch infobars, the new infobar might be using a string from the old infobar that might have been deleted.

BUG=None
TEST=Linux heap check bot should be green.

Review URL: http://codereview.chromium.org/2822032

TBR=jcivelli@chromium.org
Review URL: http://codereview.chromium.org/2879004

git-svn-id: http://src.chromium.org/svn/trunk/src@50914 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d27006f2814a887595f489802e9eb399dc0df362