Adds support for language detection on the server side. Relanding: http://codereview.chromium.org/2860037 The CLD can sometimes fail to detect the language of a page, in which case the translate server might be able to. This CL ensures the browser deals with the server side language detection properly.
Also added a new factory method to create error translate infobar delegates that does not assert when the language code is incorrect.

Also changed the enums in TranslateInfoBarDelegate2 to use the constant name style, as it is the preferred naming for enums according to the style guide.

BUG=40857
TEST=Visit http://www.saanconf.org/saanday.htm, use the context menu
to translate the page. You should get an error mentioning the
page is already in English.

Review URL: http://codereview.chromium.org/2854032

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@51288 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 68460d5ea0c0b67caa4c6481d0fb697266e10c6f