Nudge syncer thread when notifications are enabled.
Several sync integration tests were failing because the fact that
notifications were enabled was not detected by the sync clients. This
changelist adds a call to NudgeSyncer() in OnNotificationStateChange()
so that the syncer thread is made aware of the fact that notifications
have been enabled. This change caused the integration test pass rate to
increase from ~20% to 90+%.

BUG=47240
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/2814045

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@51341 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 983aefb68408d544d0aaa5678ac547d2fe423eb3