Auto-size the views version of the options dialog.
This adds support for auto-sizing tab controls, adjusts the options dialog to
use it and takes care of any fallout due to the new layout handling. Also fixes
a couple of small bugs in the views Layout() machinery  and sanitizes layouting
of options pages.

BUG=36497
TEST=unit tests in tabbed_pane_unittest.cc and grid_layout_unittest.cc, as well as checking the options dialog in any supported language.

Review URL: http://codereview.chromium.org/2812026

git-svn-id: http://src.chromium.org/svn/trunk/src@51526 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4021590e0e9faf678e84013adf0966f64bf27837