Display an info link in the blocked cookies info bubble.
The link will pop up a dialog that displays the cookies blocked and accessed by the current page.

The XIB change adds a link-styled button (cell type HyperlinkButtonCell) below the title and above the horizontal line on the bubble with the text IDS_BLOCKED_COOKIES_INFO. The button action is hooked up to showMoreInfo: in the File's Owner.

BUG=45230
TEST=none

Review URL: http://codereview.chromium.org/2799012

git-svn-id: http://src.chromium.org/svn/trunk/src@51656 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e6e576345b6200957fb61cbb578db6375e395fa2