Bookmarks: Fix position of context menu when opened using the keyboard.
After fixing the contextmenu event in WebKit we can no longer rely on the bug/hack/feature that screenX was 0 when the event was generated using the keyboard. We now keep track of keydown and keyup and if the right key was down then we know it was a context menu generated by the keyboard.

BUG= 40073
TEST=Select an bookmark item (or a folder in the bookmark tree). Press Shift+F10. The context menu should be positioned relative to the selected item.

Review URL: http://codereview.chromium.org/2845042

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@51779 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b0ae05ce5daa6be50dcaeeddc52be6cb2c7f2c6d