Allow data_packages outputs to include flattened items without rc_all outputs.
This fixes a bug where the code to flatten an html file was in
the rc_all output step.  So if a data_package output came before
the rc_all output in the grd file, we would get a build error.

Just refactor the code to flatten for both output types (but only
flatten once).

BUG=48253
TEST=build completes succesfully

Review URL: http://codereview.chromium.org/2889001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@51868 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4e34d447fdfb672bc4e66f9eb6c494f7abcb8cd6