Add the capability to run multiple proxy PAC scripts in parallel.
Refactors SingleThreadedProxyResolver into MultiThreadedProxyResolver.

New threads are created lazily on demand, up to a fixed maximum.

Note that this CL does NOT change the policy in Chrome -- it will continue to use a single thread for proxy resolving, but using the new code to do so.

BUG=11079
Review URL: http://codereview.chromium.org/2822043

git-svn-id: http://src.chromium.org/svn/trunk/src@51877 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b98394c6c20dd302fdc815671299e3ce1ba8ffaa