Avoid preconnect (and pre-resolution) when proxy MIGHT be present
This change detects if there is a chance that a proxy MIGHT
be used to perform a connection, and if so, it avoids doing
any speculative preconnections, or speculative DNS
pre-resolution.

For testing purposes, a flag is provided to avoid this
probe for proxy status.  If you enable --preconnect-despite-proxy
then preconnection will take place (as appropriate) even
if there is a chance that a proxy might be used
for connections to some hosts.

BUG=47906
r=eroman
Review URL: http://codereview.chromium.org/2871038

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@52016 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1069667cb89a14340b407bafb724417d7742975a