Avoid preconnect (and pre-resolution) when proxy MIGHT be present
This change detects if there is a chance that a proxy MIGHT
be used to perform a connection, and if so, it avoids doing
any speculative preconnections, or speculative DNS
pre-resolution.

For testing purposes, a flag is provided to avoid this
probe for proxy status.  If you enable --preconnect-despite-proxy
then preconnection will take place (as appropriate) even
if there is a chance that a proxy might be used
for connections to some hosts.

BUG=47906
r=eroman
Review URL: http://codereview.chromium.org/2871038

git-svn-id: http://src.chromium.org/svn/trunk/src@52016 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 385f800de8836a1ce6331064639b478bb029e4f8