Plumb SSL connection information into the PageInfo model.
This plumbs two bits of information into the PageInfo model (the dialog
which results from clicking on the padlock icon): whether or not we
performed SSLv3 fallback and whether or not the server supported the
renegotiation extension.

It doesn't actually do anything with this information yet (except to add
histograms of them), pending future CLs.

BUG=none
TEST=none

http://codereview.chromium.org/2943001/show

git-svn-id: http://src.chromium.org/svn/trunk/src@52079 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3b5b63985a3203d7ef93bc0ea891dbe057540f3b