[Mac] Mouse clicks in omnibox close popup.
Previously a right-click in omnibox or on a page action would have a
context menu _and_ the omnibox popup.  Clicking a page action or star
would mess up the omnibox selection.  This brings things in line with
other platforms.

Also fix to not change selection if an omnibox decoration handles the click.

BUG=40145
TEST=Clicking star with omnibox popup up should close popup.
TEST=Clicking star should not change omnibox selection.
TEST=Right-click in omnibox with popup up should close popup.

Review URL: http://codereview.chromium.org/2900006

git-svn-id: http://src.chromium.org/svn/trunk/src@52126 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e6fb040a5455a80cade9404674dbf907129a165a