[Mac] First part of Omnibox decoration refactor.  Enable ev bubble.
The views implementation of the little things floating around in the
Omnibox uses nested views which are manually laid out by the
location-bar code.  The Mac code has a variety of different
implementations for these items, with the layout distributed across
many files.  This change creates something called a "decoration" which
is similar (different name because Mac has a strong sense of what a
"view" is).  Decorations are kind of like a C++ NSCell, which the
AutocompleteTextFieldCell handles generically and the
LocationBarViewMac::Layout() code will lay out.  The overall goal is
to loosely parallel the views code for decorations and Layout(), so
that future coders can do the right thing more easily.

This CL converts the left-hand items, namely:
 - the location icon
 - the search-keyword bubble
 - the ev security bubble (new)
The Layout() function does not yet deal with trimming things to fit
when space is tight.

location_bar_view_mac_unittest.mm was always a sham, so I'm going to
stop pretending.

BUG=41998
TEST=EV bubble for secure sites.
TEST=location icon can be clicked (page info panel) and dragged.
TEST=EV bubble can be clicked (page info panel) and dragged.
TEST=keyword-search mode shows appropriately.

Review URL: http://codereview.chromium.org/2805070

git-svn-id: http://src.chromium.org/svn/trunk/src@52223 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 034016fc7407d08b4321b3ca3e508e6c9e2976c0