media: refactoring video_render_base to recycle buffers
To make recycle work, we had to define the usage scope of current frame. otherwise we are introducing tearing because we will begin to decode into the buffer before it is done by the renderer/painter/compositor.

current mechanism depends on hold reference of a copied picture. we had no that luxury if we do not copy output buffers. we had to compromise by
1. in pause() ( which is not the sense of pipeline->pause(), which is implemented by set playrate = 0 ) in filter->pause() ( or in the future flush() ), which is part of seeking ( and in the future, part of stop() too) , we had to return all the buffers to owner. we had no current buffer to display. we use NULL as current frame in this case.
2. remove black frame from render base, this is only valid for system memory based video frame, even that should we use color fill instead of color conversion and scale.
3. pause and stop has to wait for pending read (actually flush) and pending paint.
4. we only advance frame when there are two or more frames in ready queue.

Review URL: http://codereview.chromium.org/2836038

git-svn-id: http://src.chromium.org/svn/trunk/src@52398 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b1fba2a732fcb9053722d80d3ddc2cf1c5b858ec