Boot prefetch optimization for Chrome Frame (experimental)
If chrome is warmed up during a single reboot, it gets paged
in for subsequent reboots and the cold startup times essentially
look like warm times thereafter! The 'warm up' is done by
setting up a 'RunOnce' key during DLLRegisterServer of 
npchrome_frame.dll.

This works because chrome prefetch becomes part of boot
prefetch file ntosboot-b00dfaad.pf and paged in on subsequent
reboots. As long as the sytem does not undergo significant 
memory pressure those pages remain in memory and we get pretty
amazing startup times, down to about 300 ms from 1200 ms

The downside is:
- Whether chrome frame is used or not, there's a read penalty
  (1200-300 =) 900 ms for every boot.
- Heavy system memory usage after reboot will nullify the benefits
  but the user will still pay the cost.
- Overall the time saved will always be less than total time spent
  paging in chrome
- We are not sure when the chrome 'warm up' will age out from the
  boot prefetch file.

The idea here is to try this out on chrome frame dev channel
and see if it produces a significant drift in startup numbers.


Review URL: http://codereview.chromium.org/2936010

git-svn-id: http://src.chromium.org/svn/trunk/src@52489 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 49fdc46c146a4602028a88bc31eea15e21258fb4