ChromeFrame tabs would hang at times while closing. This would randomly occur if the page had an unload handler. We execute unload handlers in the WM_DESTROY message in the external tab and spin a nested loop waiting for the unload handlers to finish. This causes a deadlock at times if a windows message is dispatched to IE which is blocked in DestroyWindow.
The fix is to remove the nested loop mess from the external tab and instead send over a special automation message
to Chrome in which context we execute the unload handlers. The message contains the notification window and the
actual window message to be posted back when the unload handlers finish executing. 

The active document/activex spin a GetMessage loop waiting for this message to arrive. To ensure that we don't wait
indefinitely we have a 1 second timer and exit the loop if this timer is received.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=49132

Bug=49132
Test=Covered by existing unload event test.

Review URL: http://codereview.chromium.org/3014001

git-svn-id: http://src.chromium.org/svn/trunk/src@52523 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 49e33e6ddcf8912c826d128e84756ffe141ba8f6