Set up tests to exercise the chrome outersandbox from the nacl loader. It is similar to the renderer sandbox tests in that the test code is separated into a DLL and only loaded based on commandline flags.
Currently tests file open, process creation, and connect.

This is currently not set up for Linux. To set it up for Linux,
we need to be more careful about what tests are expected to pass,
and will need to look into zygote process for how to get the test
shared lib loaded.

BUG=39409
TEST=none

Review URL: http://codereview.chromium.org/1549046

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@52567 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 123c9b79fe83e84e31994f5ef628c1603d45984c