Mac: Fix drawing when the device's transform has a scaling part.
This currently never happens in practice, but I have a patch which uses the PaintAt IPC, which doesn't work without this fix. Without this, the clip rect is scaled down too, which means if I send a PaintAt IPC that draws the renderer at 800x600 into a 400x300 pixmap, the clip rect is 200x150, because render_view sets the scale t0 0.5 and the clip rect to 400x300.

BUG=none
TEST=Everything still paints correctly.

Review URL: http://codereview.chromium.org/3033005

git-svn-id: http://src.chromium.org/svn/trunk/src@52712 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6b3860a109cf6792106e373516f682f5693996c8