AutoFill Profiles dialog handles multi-select and delete on Mac
Changes the AutoFill dialog on Mac to support multi-select and delete of items
in the address and credit card table.  The "Edit..." button disables when multiple
items are selected.

Nib changes in AutoFillDialog.xib: (1) Enables multi-select for the table view, (2) binds
the enabled state of the "Edit..." button to the new |multipleSelected| property, (3) connects edit button to outlet.

BUG=49344
TEST=AutoFillDialogControllerTest.DeleteMultiple.

Review URL: http://codereview.chromium.org/3039014

git-svn-id: http://src.chromium.org/svn/trunk/src@52989 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 62b4a47053fb3d4996a93009484ab479c667bc06