Make ResourceDispatcherHostRequestInfo own SSLClientAuthHandler and LoginHandler references
This avoids problems where the objects disappear while the
ResourceDispatcherHostRequestInfo still holds pointers. In particular,
SSLClientAuthHandler handler adds and removes a reference to itself when the
user dismisses the dialog. ResourceDispatcherHost, however, still holds a
pointer and may attempt to notify it of a cancelled request. This stray write
may NULL out a useful field and crash.

BUG=49352
TEST=steps in bug report

Review URL: http://codereview.chromium.org/3043008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@53082 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 0a6b2d022c9bdc8642f6a8d543aaea376507f537