Fix regression that stopped properly restoring focus to tab contents. The bug in r51262 was that native_tab_contents_container_win overrides RequestFocus, so it can't be *reliably* focused just using SetFocus.
This is a simple, straightforward fix that should be merged for M6.

Afterwards, I think a better solution for the dev channel would be to get rid
of the hacky RequestFocus overrides and add a View::HasNativeFocus method
so that FocusManager can tell when it thinks a view already has focus, but the
native focus might need to be set again. Thoughts?

BUG=48917
BUG=49061
TEST=none

Review URL: http://codereview.chromium.org/3013023

git-svn-id: http://src.chromium.org/svn/trunk/src@53471 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 524a52a157c2abc4d9781be302270c90a5725463