Revert 53603 - Make the OmniboxSearchHint infobar use PAGE_ACTION_TYPE.
With this, this infobar gets the blue style to match with the mocks.
Also did a cleanup to get rid of INFO_TYPE that is not more needed.

[Rollback: Following this patch, the Chromium XP debug builder started
having persistent link failures (LINK1318 error). This is a test to see
if it fixes the problems.]

BUG=39102
TEST=trybots

Signed-off-by: Thiago Farina <tfarina@chromium.org>

Review URL: http://codereview.chromium.org/2805096

TBR=tfarina@chromium.org
Review URL: http://codereview.chromium.org/2884038

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@53631 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 893e3e083fe8d77949860cdb1af7fc14a371651e