Revert 53849 - Implement most of the startup page controls in DOMUI prefs
Hooks up display of startup pages, removal of pages, and setting the startup pages to the currently open tabs.
Uses a select for now, but will eventually be replaced with a List so the favicons can be displayed.

BUG=48713
TEST=Startup pages can be viewed, removed, and set to current in DOMUI prefs.

Review URL: http://codereview.chromium.org/3046025

TBR=stuartmorgan@chromium.org
Review URL: http://codereview.chromium.org/3058018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@53857 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 00281662f6b28c7c0b84d76761f8e294f2b603e2