Implement accurate seeking by decoding audio/video until we reach our desired timestamp.
This fixes a host of seeking, synchronization and layout test issues but does incur a CPU performance hit while seeking.  In the worst case a seek could take quite some time if the content was high resolution H.264 and we attempted to seek to right before a keyframe.

We'll see how people like the new behaviour and back this change out if needed.

BUG=48984
TEST=layout tests, media_unittests

Review URL: http://codereview.chromium.org/3032030

git-svn-id: http://src.chromium.org/svn/trunk/src@53872 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b193af2aee8c186f72fd0c218b2b25c8572327c9