Fix glitch in rendering of the stroke at the bottom of the browser toolbar.
The code that rendered this line was doing weird things (adding/removing kClientEdgeThickness from the edges?!) so I removed this and used toolbar_bounds.right() instead of toolbar_bounds.width() which is correct given that toolbar_bounds.x() > 0.

I also got rid of the use of MirroredLeftPointForRect since the flipped version of a horizontal line is... the same horizontal line.

I have also updated the glass frame not to subtract a pixel from the right edge which looks glitchy as well.

http://crbug.com/48134
TEST=see bug

Review URL: http://codereview.chromium.org/3056022

git-svn-id: http://src.chromium.org/svn/trunk/src@53998 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: addbce358c0883e97bc862232ac8b92fd1aad023