Add an extractor for DOM features to be used for client side phishing detection.
PhishingDOMFeatureExtractor iterates over the page elements and computes a
number of features.  To avoid blocking the renderer for too long, the extractor
may run in several chunks of works, posting a task to continue processing if
necessary.

This CL only includes the feature extraction itself.  I will add the logic to
cap the time per iteration in a follow-up CL.

BUG=none
TEST=PhishingDOMFeatureExtractorTest

Review URL: http://codereview.chromium.org/2878046

git-svn-id: http://src.chromium.org/svn/trunk/src@54082 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 71ab8f7892dec24c5557222db80e0e30c7a1dcde