Give a better error when extension unpacking can't be done.
The sandbox can not allow file access to paths that contain reparse points.  Chrome tries to find a reparse point free path to the directory which will be sandboxed for extension unpacking.  However, there are cases where there is no such path.  See the bug for an example.  In this case, give a decent error message.

BUG=49530
TEST=Manually created a partition mounted in C:/mnt, mounted as drive letter, and mounted only under /Devices/HardDisk3/, tried unpacking extensions in each.

Review URL: http://codereview.chromium.org/3060026

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@54327 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 40ea93d42b3962876bfa00e806546aa6d8881a2f