Fix a bug with detecting window sizes that match the monitor size.
The previous code was comparing the window size to the screen
size, but that doesn't work properly for twinview or other cases
where there are multiple monitors that make up a single screen.
Instead, compare the window size against all monitors (since we
don't know which monitor the window is going to be placed on)
and resize if it matches any monitor size.

Also fix a bug in BoundsMatchMonitorSize where we should just be
comparing sizes, not rectangle positions.

BUG=32347

Review URL: http://codereview.chromium.org/3060032

git-svn-id: http://src.chromium.org/svn/trunk/src@54332 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0fe8da87f68c0137d282489aaf0917a0b70a922a