Add the actual data being read to the OnBytesRead callback, take two.
This change was originally committed as
http://src.chromium.org/viewvc/chrome?view=rev&revision=53178
then rolled back as
http://src.chromium.org/viewvc/chrome?view=rev&revision=53416
due to a breakage in FLAKY_UnknownSize download test.
FLAKY_UnknownSize depended on code with a bug that was exposed by this change. That bug has since been fixed and committed as
http://src.chromium.org/viewvc/chrome?view=rev&revision=53876

So this change is ready for another round of review. It has not changed at all since being committed.

Contributed by: bmcquade@google.com

BUG=48192
TEST=Added new unit tests, ran net_unittests.

Review URL: http://codereview.chromium.org/3010037

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@54448 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 79268c34d39b0cee1ca86d101c097145378089af