Make PhishingDOMFeatureExtractorTest heapcheck suppressions more specific.
The root cause of these leaks is that the IPC::ChannelProxy is never freed.
This happens because the ChildThread destructor intentionally detaches the
channel from its message loop prior to the channel being destroyed, so the
ChannelProxy's self-reference is not released when the channel closes.  In
practice, the leak should only occur just before a renderer is shut down, so
it shouldn't be a problem.

BUG=50598
TEST=tools/heapcheck/chrome_tests.sh -t unit --gtest_filter=PhishingDOM*

Review URL: http://codereview.chromium.org/3035054

git-svn-id: http://src.chromium.org/svn/trunk/src@55152 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a18655658051e2cb955a3a6ec81342806c73960c