Make PhishingDOMFeatureExtractorTest heapcheck suppressions more specific.
The root cause of these leaks is that the IPC::ChannelProxy is never freed.
This happens because the ChildThread destructor intentionally detaches the
channel from its message loop prior to the channel being destroyed, so the
ChannelProxy's self-reference is not released when the channel closes.  In
practice, the leak should only occur just before a renderer is shut down, so
it shouldn't be a problem.

BUG=50598
TEST=tools/heapcheck/chrome_tests.sh -t unit --gtest_filter=PhishingDOM*

Review URL: http://codereview.chromium.org/3035054

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@55152 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ab6cd09c17e6a1427457c56515996eb804dfd233