The Selenium window.open target blank test fails in ChromeFrame as the navigation is treated as a non local top level navigation and is routed to IE thus causing the test to fail.
The test calls window.open with an empty url. The renderer logic which detects whether the navigation is to be routed
to the external host matches the origin of the frame and the url and as they differ ends up routing the navigation to
the host.

Fix is to allow only http/https navigations to be routed to the host. All other navigations stay within Webkit.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=51412

Bug=51412

Review URL: http://codereview.chromium.org/3044053

git-svn-id: http://src.chromium.org/svn/trunk/src@55300 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3c452ae91707d1bbfdf10d1740089aca0ee27514