Add remoting to wrench menu.
This change adds remoting item to wrench menu.
The change is under flag enable-remoting, so end users will not see it.

The CL is cleaned up. It was compiling and working. I added the check for kEnableRemoting and resynched and it is no longer compiling.

Submitted for: pranavk@chromium.org
Reviewed: http://codereview.chromium.org/3075024/show

BUG=50242
TEST=NONE

Review URL: http://codereview.chromium.org/3080038

git-svn-id: http://src.chromium.org/svn/trunk/src@55311 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a7e943d7a542e3fcb55a658cf336d01e8e808d3e