Revert 55311 - Add remoting to wrench menu.
This change adds remoting item to wrench menu.
The change is under flag enable-remoting, so end users will not see it.

The CL is cleaned up. It was compiling and working. I added the check for kEnableRemoting and resynched and it is no longer compiling.

Submitted for: pranavk@chromium.org
Reviewed: http://codereview.chromium.org/3075024/show

BUG=50242
TEST=NONE

Review URL: http://codereview.chromium.org/3080038

TBR=hclam@chromium.org

git-svn-id: http://src.chromium.org/svn/trunk/src@55316 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c9395f1d80642745055d538ceebe8e8c298b625b