(repair leak) Protect instance after commiting to use network
The code was incorrectly protecting an instance of Preconnect
before deciding it would actually perform the async network
calls, and need the instance to stay alive for a callback.
The leak was caused when we exited the function, did not
make the asnyc call, but still incref'ed the instance
expecting a callback to come and decref it.

BUG=51133
r=wtc
Review URL: http://codereview.chromium.org/3086023

git-svn-id: http://src.chromium.org/svn/trunk/src@55435 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f923ab4517375cbf56d3a52d14a6d0818a8b1ba1