Use SSL/TCP port for notifications in sync tests.
The sync integration tests start off by using the XMPP port (5222) to connect to the sync notification server. Since this port is blocked on the chromium builders, the tests wait for upto a minute before they failover to the SSL/TCP port (443), by which time the tests time out and fail.

This checkin adds a switch that, when enabled, causes the tests to try the SSL/TCP port before the XMPP port.

BUG=50595
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/3050036

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@55599 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ef649c8de5b95bd488922473c8ca529afaae9bfb