GTK: minor MenuBarHelper fix.
If one of the buttons is not realized, gtk_widget_translate_coordinates() will fail on it. Instead of aborting, just skip to the next button. This can happen when opening a new window with a bookmark bar that doesn't have an overflow chevron (when the chevron has not been shown, it will be unrealized).

BUG=51810
TESt=see bug + above description

Review URL: http://codereview.chromium.org/3175006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@55799 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: fef25cef03899418a91757ae5c8c515a8018be37