Allow per-plugin content settings.
Run with --enable-resource-content-settings and --enable-click-to-play to enable.

TabSpecificContentSettings now keeps track of which resources were blocked. The content setting bubble displays those resource, and selecting the "allow radio buttons adds exceptions for these resources.

The infobar for non-sandboxed plug-ins now also shows a button to always allow the blocked plug-in on that site.

Screenshot of the plugin bubble: ﻿http://imgur.com/6npqv.png

Because we need to change the title and radio button labels depending on whether we track specific plug-ins, we copy them from the ContentSettingBubbleModel.

XIB changes: Add a text field for the blocked plug-ins to ContentBlockedPlugins.xib, hooked up to the |blockedResourcesField_| outlet.

BUG=39252,38432
TEST=manual

Review URL: http://codereview.chromium.org/2873104

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@55888 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5949b2806e89f233e67748fb047b7cf5c619cb3f