Run tests faster.
I noticed that net_unittests was spending a lot of time doing nothing.
This CL changes the timeouts so that we can startup and shutdown test
servers faster.

(This is a reland. Originally r54579, reverted in r54592. This is
relanding unmodified as I don't believe that it was mistakenly
fingered for build issues at the time. The real problem was r54576,
reverted in r54595.)

BUG=none
TEST=net_unittests

http://codereview.chromium.org/3057021/show

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@56024 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: db22a0617e8992089d8219d0030e394b8c5af023