Relanding r55510:
This CL originally caused crashes on the bots. I tracked the problem down to a bug in ANGLE that I fixed upstream:

http://codereview.appspot.com/1956042/show

This fix was committed upstream and I rolled the Chrome DEPS so the bot crash issue should be resolved.



Original comment for r55510:

Prelaunch GPU process after 10 seconds.

This will happen if the --prelaunch-gpu-process command line switch is present
or for 5% of users in a field trial.

This is so we can collect some stats of GPU capabilities.


TEST=try
BUG=none

Review URL: http://codereview.chromium.org/3179010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@56239 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f589b478d556675cbf5ba20ba3f0642db9c02171