DnsRRResolver: hoist inner class and callback on MessageLoop.
This change moves the Response class to the top level to save other
files from having to #include the header in order to get a pointer to
a Response object.

This change also modifies the semantics of resolution: the callback
will now be made of the origin MessageLoop, rather than a random
thread.

TEST=net_unittests
BUG=none

http://codereview.chromium.org/3110015/show

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@56342 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4e15adfcdb86c2d723558291412327377c795bc4