Acquire lock before iterating over list of field trials
The itteration is performed in the browser process
when the list of field-trial settings is passed to
the renderer.  I failed to acquire the lock prior
to the itteration, and it appears that some code
is deciding on a field trial VERY late in browser
startup and racing against this render-spawn
operation.  The code in FieldTrial is wrong
(but fixed here), but I'll still need to
track down the racing thread that is performing
such a late FieldTrial definition (the results
of which will not be passed to some renderers).

BUG=49409
R=huanr
Review URL: http://codereview.chromium.org/3134015

git-svn-id: http://src.chromium.org/svn/trunk/src@56360 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 8c1f49fdd658affc0d70a78ba6ab3785f9378fca