Acquire lock before iterating over list of field trials
The itteration is performed in the browser process
when the list of field-trial settings is passed to
the renderer.  I failed to acquire the lock prior
to the itteration, and it appears that some code
is deciding on a field trial VERY late in browser
startup and racing against this render-spawn
operation.  The code in FieldTrial is wrong
(but fixed here), but I'll still need to
track down the racing thread that is performing
such a late FieldTrial definition (the results
of which will not be passed to some renderers).

BUG=49409
R=huanr
Review URL: http://codereview.chromium.org/3134015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@56360 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6f5b175d823d7f122e8b45d76c8a8c576fc4e4a4