Handling empty ScopedDirLookup username in HandleDirectoryManagerEvent().
The constructor of syncable::ScopedDirLookup takes the username of the
account being synced. There are times when the username is blank,
resulting in a DCHECK being fired for the value of good() within the method
SyncManager::SyncInternal::InitialSyncEndedForAllEnabledTypes().

This patch makes sure that a ScopedDirLookup with a blank username is
never created by adding a check in the method
SyncManager::SyncInternal::HandleDirectoryManagerEvent().

BUG=51820
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/3184009

git-svn-id: http://src.chromium.org/svn/trunk/src@56403 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 827e5497d856ba7170f235ab18f930dfd63ddc94