Clean up the client certificate filtering code on Linux
Instead of filtering over all certificates and then picking the ones suitable
for client auth, NSS provides a function that does this in one go. In addition,
delay checking for the private key until we actually send the certificate.
While this means we may potentially offer a client certificate with no matching
private key, such a store should not exist. This avoids requiring us
authenticate to every certificate offered. In addition, it delays the password
dialog, when we implement it, until after the user has designated that they are
interested in a certificate.

This mirrors Firefox's use of NSS.

R=wtc
BUG=none
TEST=linux ssl client auth works

Review URL: http://codereview.chromium.org/3177018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@56404 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 37e429b19fc98065dfe4221cb4caf11b137ba13b