Disable autocomplete for search keywords.
We have custom UI for search keywords anyway, so in practice a user would never see the autocompleted url.  But, we had been computing the index incorrectly in some cases, as we trim off "http://" and "www." to compute the keyword.  Since we don't use it anyway, simplify the logic and get rid of it :)

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3151020

git-svn-id: http://src.chromium.org/svn/trunk/src@56445 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2d673985e7e95c000802f77058e29df8697b07c8