Allow data URLs to load chrome-extension:// resources. Basic notifications use data: URLs internally so this was required to support that.
Also rejigger the logic in this area again to clarify and fix
an issue where the incognito window check did not have high
enough priority.

Also add back a null-pointer check that was previously in this code and is apparently necessary with a comment.

BUG=52313,52374

Review URL: http://codereview.chromium.org/3159019

git-svn-id: http://src.chromium.org/svn/trunk/src@56470 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1d4a8f29494fdcaab483458ccfb28a9c8847fb5a