Disable outdated plug-ins on a per-plugin basis.
Change the automatic disabling of outdated plugins (with --disable-outdated-plugins) so that a single plugin is disabled if it is out of date instead of the whole group.

This makes it possible to fall back on a newer version of a plugin that would otherwise have lower priority.

BUG=47731,51728
TEST=Install an old version and a new version of a plugin, run with --disable-outdated-plugins. The old version should be disabled, and the effective version of the plugin group should be the new one.

Review URL: http://codereview.chromium.org/3117009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@56494 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f6da538dede3214a242c2ee9bc49968d8c94e01a