Fix a few test failures when landing http://trac.webkit.org/changeset/57178. http://trac.webkit.org/changeset/57178 broke some tests in browser_test and ui_test.
In the test RedirectTest.ClientCancelled, it needs a user-initiated event to trigger the redirect, now it uses "javaScript:click()". When landing r57178, the call of window.open which is inside the call of javaScript:click() was treated as non user-initiated, so the in-page location change was treated as client redirect and the redirect was recoreded, that is why this test was failed when landing r57178. (Please refer to the logic in FrameLoaderClientImpl::dispatchDidNavigateWithinPage.)

In the tests AppApiTest.AppProcess, ExtensionBrowserTest.WindowOpenExtension and ExtensionBrowserTest.WindowOpenInvalidExtension, they assume the new tabs opened by window.open in current window. But when landing r57178, since those calls of window.open were treated as non user-initiated, the disposition type of new tabs were changed to Popup, which caused a few new tabs were created instead of a few new tabs in current window (Please refer to RenderView::show), which cause those tests were failed when landing r57178.), that is why those tests were failed when landing r57178.

BUG=17655
TEST=RedirectTest.ClientCancelled, AppApiTest.AppProcess, ExtensionBrowserTest.WindowOpenExtension, ExtensionBrowserTest.WindowOpenInvalidExtension

Review URL: http://codereview.chromium.org/3136019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@56834 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1ebbced3047dc94ffbccd7773d263efb38dbec5d