Remove attempt to be smart about where to open navigations targetting app tabs. I futzed with this a bit to try and integrate installed apps extents, but it is complex.
Also, I looked over the original bug where this code was
added: crbug.com/29281. The issue there was that when you
click on a bookmark (presumably in the bookmark bar) the
pinned tab is unexpectedly overwritten. I can see that
complaint, but there is also the use case of clicking
a bookmark and intending it to overwrite the pinned tab.

In summary, I kinda feel like the expectations about
how navigation should work are too ingrained and we
shouldn't be meddling.

BUG=52680

Review URL: http://codereview.chromium.org/3161037

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@57002 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 09a33d5629ae539336957326671229a7c355722e