Mac: Correctly show/hide compositor on navigations.
The problem was that unlike plugins, the compositor surface is not destroyed on navigation, only a variable is_gpu_rendering_active is updated. The fix is to tell the RenderWidgetHostView if this variable changed and then to adapt the visibility of the view rendering the compositing layer accordingly.

Also, destroy its AcceleratedSurfaceContainer when the compositor is destroyed.

BUG=52134,51748
TEST=go to http://webkit.org/blog/386/3d-transforms/, then hit back. page should redraw on back button. go forward, should work too.

Review URL: http://codereview.chromium.org/3132038

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@57135 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: aa224da0bb3d423e59c865a53525748fa449275f