Adds a speech input info bubble.
I have not implemented the Mac/Linux part yet, will do so in a subsequent CL (does InfoBubble exist for Mac?).

This info bubble remains visible on screen until recognition ends or the user cancels it.
It also tracks the html element which requested speech recognition if it changes position
on screen (e.g. user scrolls the page or moves the browser window).

To get this working, I extended InfoBubble to query the delegate if it can be closed when
another window gets activated, as well as to reposition it when required.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/3133034

git-svn-id: http://src.chromium.org/svn/trunk/src@57199 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a1eeaf2e9dab49ab63129a503423b56b3c6be2c9