Mac: Well-behaved accelerated plugins, actual fix
Configure the opengl view to draw below the window, draw the view with NSClearColor so that's transparent, and make the window non-opaque while plugins are showing, so that the opengl surface can show through.

This is a bit slower than previously without this patch, but it's about as fast as it was when we used CoreAnimation to show IOSurfaces.

Parts of this code are from Simon Fraser's MacTierra code (with permission).

BUG=44087,51748
TEST=
* Go to youtube. Findbar should show up on top of video. Go fullscreen. Tab overlay should show up on top of video.
* Go to youtube, drag tab with video out and to a new window. Video should continue playing.
* Install https://chrome.google.com/extensions/detail/bdnkaenpadjoldiddfdidinjmjeagaji , click browser action. Should still work.
* Install a transparent theme such as https://tools.google.com/chrome/intl/cy/themes/theme_at_mecko.html or http://chromium.googlecode.com/issues/attachment?aid=3778755830122130591&name=theme-bug.crx&token=27908ca7e699a023c8446657b24c4696 . Browser window shouldn't become transparent while videos are playing.

Review URL: http://codereview.chromium.org/3176027

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@57201 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1adb91c25bd546f872c3112fd912914d983c3e1e