Fix a crash where we are checking IsConnected().  If you look into the implementation of IsConnected() on windows, you'll find that it can synchronously check if the socket is connected.  This makes it so that a CHECK (or DCHECK) on this attribute is dependent on the peer.  The peer could send us a TCP-FIN at any time - even 1 nanosecond before our call to IsConnected().  We can only use CHECK(!IsConnected()) to check before we connect, we simply can't use it at runtime.
I searched the code and found 4 other instances of this broken check.  Most
were DCHECKs, which is good, but I've removed them just the same.

Turns out our MockSockets have a mechanism for simulating this error, but it
wasn't tested nor plumbed through the MockSSLClientSocket.  In the process
I added a couple of basic proxy test cases, which may be slightly redundant
with other test cases that are more advanced.  But they seem like a good
idea to cover anyway.  Added 6 tests in total.


BUG=53099
TEST=ProxyTunnelGetHangup, RecycleDeadSSLSocket

Review URL: http://codereview.chromium.org/3107036

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@57295 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8f08117166723655ffe7fa23e5c597128ff38294