Revert 57386:
The interval test started failing after we rolled back to a previous version of Webkit. This code
does not seem to support it. 

Original description: Move the chromium.Interval to chrome.Interval as part of the benchmarking
extension.  This means that users will need to use --enable-benchmarking
to access it.  It really shouldn't be part of the exposed API.



BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3126029

TBR=mbelshe@chromium.org
Review URL: http://codereview.chromium.org/3186033

git-svn-id: http://src.chromium.org/svn/trunk/src@57459 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3c0f48822c1453a6e3d3b01ae10da9f5dffe8dd6