Add hidden component app for web store.
This patch adds a component which does not appear on the NTP, but does enforce that gallery urls are isolated in their own process. This is in anticipation of exposing extension api bindings to the gallery so it can install/detect/uninstall extensions & apps.

Note that this patch works correctly with the --apps-gallery-url. If specified, the replacement url will be inserted into the component app's extent

BUG=27431

Review URL: http://codereview.chromium.org/3116040

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@57460 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 0f416a5b84f458289aa9d57a8f8fcf9ab9d36aa6