Implement OSX version for "Start on Login" for background-enabled extensions.
When user installs first bacground-enabled extension we set Chrome to be a "Login Item", and when user uninstalls the last one, we remove Chrome from the "Login Items List".

Login Items are the apps that are listed in Systempreferences->Accounts->Login Items, or which have a "Open on Login" check for their Dock icon.

Since we manipulate the user-faced UI item here, we try to do it minimally (only on first installed and last uninstalled background app) with understanding that when we have UI for install/uninstall we'll convert this into more visible, cancellable suggestion on that UI.

Using the same bit that user can manipulate separately and perhaps even knows about is beneficial since it provides for unified place to "start Chrome on Login" and reduces confusion.

The patch also removes SetKeepAlive/EndKeepAlive calls for OSX since on OSX they are already used properly to keep Chrome alive when all windows are closed (in other words, on Mac there is only one behavior that also fits what is needed for background apps)

Implementation for Login Items List access uses LSSharedFileList API that is availabel starting 10.5, which I think is our minimal supported version of OSX.

Also added a check in BackgroundModeManager::RemoveStatusTrayIcon() which could crash on OSX because there is no icon_tray_ yet.

Review URL: http://codereview.chromium.org/3163029

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@57569 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: bd97c24112b63e69dfc0a4f0b22be6518d478373