Re-land. Fixed previous build error.
Implement OSX version for "Start on Login" for background-enabled extensions.

When user installs first background-enabled extension we set Chrome to be a hidden "Login Item", and when user uninstalls the last one, we remove Chrome from the "Login Items List".

Login Items are the apps that are listed in Systempreferences->Accounts->Login Items, or which have a "Open on Login" check for their Dock icon.

Since we manipulate the user-faced UI item here, we try to do it in a way that will likely not interfere with user settings. Most typically, users do not use 'hidden' property, so if we ever encounter setting w/o this property, we won't override. Also there is a hidden preference used to remember the fact that Chrome changed the settings earlier - to avoid override on uninstall of the last persistent app in opposite case.

Using the same bit that user can manipulate separately and perhaps even knows about is beneficial since it provides for unified place to "start Chrome on Login" and reduces confusion.

Implementation for Login Items List access uses LSSharedFileList API that is available starting 10.5, which I think is our minimal supported version of OSX.

BUG=43382
TEST=none

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=57569
Reverted, trying to re-land.

Review URL: http://codereview.chromium.org/3163029

git-svn-id: http://src.chromium.org/svn/trunk/src@57614 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4efd17b4fb995b9e9a5354357fc253be18a2d889