Displays a speech input UI bubble during speech recognition.
The webkit code passes in display rect of the input element requesting speech input
and we create a UI bubble pointing at this element when starting speech input. The
user can click outside the bubble to close it, which aborts recognition if we were
still recording audio. The user can also click the cancel link in the bubble which
aborts recognition irrespective of what state it is in now.

Added a SpeechInputBubbleController class to take care of marshalling the requests
between the speech input code in IO thread and the UI bubble in the UI thread. Also
added a unit test for this class.

BUG=none
TEST=unit_tests --gtest_filter=SpeechInputBubbleControllerTest.*

Review URL: http://codereview.chromium.org/3156048

git-svn-id: http://src.chromium.org/svn/trunk/src@57666 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b209a806d28f2880156a9093abcd1489d4bf6557