Fix DeviceOrientationProviderTest.
There was a race that could potentially result in a reference to Provider being
left after a test finished, causing the next test to fail. The test must not
assume that the orientation update notification is posted on the message loop
right away, but rather wait for it.

I was not able to reproduce the bug, but I believe this should fix it.

BUG=53468
TEST=unit_tests --gtest_filter=DeviceOrientationProviderTest*

Review URL: http://codereview.chromium.org/3231003

git-svn-id: http://src.chromium.org/svn/trunk/src@57719 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a831c5d56e93ab5ad71ce53d08b13b7f146146a0