Add way to create a texture in parent's context and copy into it
For canvas we need a way to create multiple offscreen render targets within
the same OpenGL context and expose them all to the compositor, which exists
as a parent context.  ggl currently provides a single framebuffer and texture
pair per offscreen context and implements swapBuffers() to copy to the parent
texture.  This generalizes that to let clients create as many parent textures
as necessary and do the equivalent of swapBuffers() for each without the
service side needing to keep track of the mapping.

We'll want to change this in the future when we start using glBlitFramebuffer
to support multisampling.  I think in that world we'll need a way to bind a
texture in the parent's context to the DRAW_FRAMEBUFFER context.  For now,
I think this is OK.

TEST=open a page with multiple accelerated canvases in a shared context
BUG=52684

Review URL: http://codereview.chromium.org/3211005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@57761 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 15ecb85585e516769dbfeb146f034ee7445cce5b