Jingle_glue bugfixes.
Fixed Closed() methods in JingleChannel and JingleClient so that they are
not blocking and guarantee that callback is not called afterwards.
Changed JingleThread shutdown mechanism so that the thread doesn't stop until
all tasks are finished.
JingleClient now uses LocalTunnelSessionClient that allows local connections.

BUG=52889
TEST=unittests

Review URL: http://codereview.chromium.org/3167047

git-svn-id: http://src.chromium.org/svn/trunk/src@57770 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c4716e56cdbc322fbbb4e0a59bfc5c219ec96268