Add in some infrastructure to make tracing of logical requests broken over async callbacks easier.
Create a concept of a TraceContext that holds a current Tracer which can be
used to create a timeseries sequence of events.

This implementation extends the functionality of RunnableMethod and
NewRunnableMethod() to create a parallel TracedMethod, and NewTracedMethod().
These new Traced methods methods and types know how to propogate the current
TraceContext to create a logical sequence of annotations through the servicing
of one logical action.

Currently, the code is controlled via define macro which changes
NewTracedMethod back into NewRunnableMethod if the hook is disabled.

BUG=52883
TEST=compiles, and client can connect to host.

Review URL: http://codereview.chromium.org/3169047

git-svn-id: http://src.chromium.org/svn/trunk/src@57992 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2006d05b9798fc29e149518e83e41295ff603a0f