Cleanup: make more PrefMember methods const.
This allows the users of PrefService to make their accessors
that refer to PrefMember members const too.

We use mutable for some PrefMember members, when they
do not reflect the outside-visible state. Those members
are used to do lazy loading of the preference values.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3271004

git-svn-id: http://src.chromium.org/svn/trunk/src@58086 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ff7e378fb4e79e1d44f7d2b290a6e6ce6d17ab7e