Pull seccomp-sandbox in via DEPS rather than using an in-tree copy
This means changes to the sandbox won't have to be committed twice, to 
both trees. 

This is a retry of r57921, which was committed with git-svn and failed 
to remove the "seccomp" directory. This caused problems when trying 
to "svn checkout" to the same location, and the change was reverted. 
This time I will use SVN to commit the change. 

BUG=none 
TEST=smoke test of running chromium with --enable-seccomp-sandbox 
Review URL: http://codereview.chromium.org/3225010

git-svn-id: http://src.chromium.org/svn/trunk/src@58184 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4db701069f477e5f8a33efd59ef1931653223aac