Complete initializing a GPU video decoder in a GLES2 context
In this patch a ggl::Context is connected to the GpuVideoDecoderHost in the
renderer process. In the GPU process the GpuVideoDecoder is connected to a
gles2::GLES2Decoder.

These changes will be used in the future to switch context before issuing
video decode commands. This is also needed by the GPU process to translate
a client texture ID to a service texture ID in the GPU video decoder.

BUG=53714
TEST=none

Review URL: http://codereview.chromium.org/3266008

git-svn-id: http://src.chromium.org/svn/trunk/src@58518 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c241efde61c30f9e207ee17e64d2e101a97e7791