Add a term feature extractor for client-side phishing detection.
This class creates features for n-grams in the page text that appear in the
phishing classification model.  It will eventually operate on the plain text
that is extracted by RenderView::CaptureText().

To make it harder for phishers to enumerate the terms in the classification
model, they will be supplied as SHA-256 hashes rather than plain text. The
term feature extractor hashes the words in the document in order to check
whether they match the model.  Since this is potentially expensive, the term
feature extractor limits how long it will run on each iteration, similar to
the PhishingDOMFeatureExtractor.

TEST=PhishingTermFeatureExtractorTest
BUG=none

Review URL: http://codereview.chromium.org/3214002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@58537 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d87552c613829fc65bcd19cc8119d525802d7fc0