[CHrome OS] Provides nicer abstractions around ownership API
There are two categories of operations that can be performed on the Chrome OS owner-signed settings store:
1) doing stuff to the whitelist (adding/removing/checking)
2) Storing/Retrieving arbitrary name=value pairs

Unfortunately, it is currently a limitation that only one of a each category can be in-flight at a time.  I've filed an issue on me to remove that restriction.

The pattern of use here is that the caller instantiates some subclass of SignedSettings by calling one of the create methods.  Then, call Execute() on this object from the UI thread.  It'll go off and do work (on the FILE thread and over DBus), and then call the appropriate method of the Delegate you passed in -- again, on the UI thread.

BUG=chromium-os:4488
TEST=Unit tests

Review URL: http://codereview.chromium.org/3345002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@58546 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 79fb66cd72a0d8787d83373f52d96c075bb63255