Adds the following to CancelableRequestConsumer:
GetClientDataForCurrentRequest: The particular use case I need this
for is that I'm using the same CancelableRequestConsumer for multiple
sources so that I don't know whic CancelableRequestProvider to pass in
to GetClientData. None-the-less it seems a lot easier to use this when
always get the provider and handle.

GetFirstHandleForClientData: For this one I know I want to cancel a
request for a given client data but don't know the associated handle.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3311012

git-svn-id: http://src.chromium.org/svn/trunk/src@58759 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7e38edb19542c73519921b7d6bb57a0249327909