Mac: Try to get tabpose unit tests working again.
The original problem was that the test plays an animation that outlives the test. When the animation is done, the tab strip model observer tries to unregister itself, but the registry died when the test exited already.

My faulty fix was to reset the observer as soon as the end animation starts. This has the problem that the end animation will try to access stale objects: TabContentss are destroyed when the test exits, but the exit animation still tries to access them when drawing ThumbnailLayers (because the observer didn't get a notification that these were dead, since it was unregistered).

This (hopefully better) fix now unregisters the observer directly before the tab strip goes away.

BUG= 54323,53893,54342
TEST=unit tests pass, valgrind stays green

Review URL: http://codereview.chromium.org/3304014

git-svn-id: http://src.chromium.org/svn/trunk/src@58945 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3e7d671ac75a186fa74dab4f80c3f61d09647779