Use text-overflow: ellipsis for long words in sync set-up UI.
In r57340, the datatype selection section of sync set-up UI got wider, but
not wide enough to avoid truncation in all the locales. When a word 
is too long to fit, it's truncated with 'tooltip' showing the full-word.

This CL adds 'text-overflow: ellipsis' and other necessary CSS properties 
to 'label' elements so that long words that do not fit are ellipsized instead of
truncated.

While I'm at it, I fixed two 'typos'. One is 'i8n' (note missing '1') and
the other is the misplaced label for the radio-button for 
'Choose what to sync'. The raido-button and the corresponding label is
separated by 'div' tag (chooseDataTypesBody) leading to a strange line wrapping
in the sync setup UI.

BUG=52789, 53494
TEST=1. Open sync setup UI in French, Italian, Brazilliant Portugest (pt-BR) and see if the translation for 'autofill' is ellipsized rather than truncated.
2. Open sync setup UI and 'Choose what to sync:' is on the same line as the corresponding radio button.

Review URL: http://codereview.chromium.org/3142042

git-svn-id: http://src.chromium.org/svn/trunk/src@58979 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 093ce65f8d3e57cbbab2d3164c0c23f2a02b3109