Fix 'about:terms' for Swahili and Amharic.
terms_sw.html was taken from the download page. We don't yet have
the EULA page in Amharic. So, for Amharic, we use terms_en.html
until it's available.

BUG=54999
TEST=Build Google Chrome and go to about:terms. In Swahili Chrome, the page should come up in Swahili. In Amharic Chrome,
it should come up in English for now. Once the translation is available, it'll be replaced.
TBR=mal
Review URL: http://codereview.chromium.org/3368009

git-svn-id: http://src.chromium.org/svn/trunk/src@59009 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 9a77f4351254c3a39b026eadfd32ca0cdb2baf55