Fix 'about:terms' for Swahili and Amharic.
terms_sw.html was taken from the download page. We don't yet have
the EULA page in Amharic. So, for Amharic, we use terms_en.html
until it's available.

BUG=54999
TEST=Build Google Chrome and go to about:terms. In Swahili Chrome, the page should come up in Swahili. In Amharic Chrome,
it should come up in English for now. Once the translation is available, it'll be replaced.
TBR=mal
Review URL: http://codereview.chromium.org/3368009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@59009 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 21ca5971edbe04c4f34244e2193bba57ef415adb