Revert 59019 - This broke the EncryptExportDecrypt unittest.
Original description:

Add a scorer class for client-side phishing detection.

The Scorer class loads a machine learning model and lets you compute a
phishing score for a set of features that were extracted from the current
page.  The phishing score corresponds to the probability that the page is
phishing.  The features are extracted by the various feature extractor classes
that live in the same folder.  The current format of the model is a
protocol buffer (see: client_mode.proto).

BUG=NONE
TEST=PhishingScorerTest

Review URL: http://codereview.chromium.org/3363004

TBR=noelutz@google.com
Review URL: http://codereview.chromium.org/3317017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@59034 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f1529d6ec6f77fb640770321a9b37fd1f09b3f33