Change the toast text and make sure errors are written to HKLM too (not just success messages).
The toast text changes are listed in the bug; basically we
have 4 experiments that differ only in the text we display
as a headline.

I also fixed an issue raised by QA with error codes being
written to HKCU as opposed to HKLM. This should fix that
issue.

BUG=http://crbug.com/54835, http://crbug.com/44378
TEST=Requires manual testing by QA.

Review URL: http://codereview.chromium.org/3341026

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@59102 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5a28b246588679258ce70063b9c12cad4ab14132