Add a confirmation prompt to app uninstallation on the ntp.
The uninstall dialog used to say:

[===============================x]
|Confirm Uninstallation          |
|--------------------------------|
|<b>Uninstall Foo Extension?</b> |
|                                |
|Are you sure you want to        |
|uninstall this extension?       |
|                                |
|================================|

We don't have the bottom string with the word 'app', and I
realized it is somewhat extraneous anyway. So just removed it
in all cases and de-bolded the text above. Looks much better.

Also, fixed a bug where the icon we display in the GTK
dialogs is too big by doing image resizing in
ExtensionInstallUI.

BUG=54874
TEST=Uninstall app and extension. Both should have a prompt,
and UI layout should be the same.

Review URL: http://codereview.chromium.org/3332016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@59157 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 77105e84fa7fbf94eb7c93925da10a9ff13a6c18