net: disable renegotiation checks.
Although we have a handle on the problems of corporate MITM proxies,
users are now reporting the anti-virus programs are choosing to
implement their virus scanning by MITM attacking local SSL/TLS
connections.

It only appears to be some anti-virus programs in some configurations,
but we don't really have a good idea of the scope of the issue so this
patch disables the renegotiation checks (for merging to the stable
branch) while we work on a better solution.

BUG=55410
TEST=none

http://codereview.chromium.org/3340024/show

git-svn-id: http://src.chromium.org/svn/trunk/src@59241 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 211944d94c76a1826ca898ba2405aa02e99c3e5d