Extend speech input bubble in GTK to display error messages with try-again and cancel buttons. The SpeechInputBubbleGtk object now lives longer than the actual info bubble window, and allows the caller to create a bubble on screen or update with an error message when needed.
When recording speech, the layout of controls are (vertical):
- Label ('Speak now')
- Icon (Mic or wait)
- Button Bar (Horizontal, 1 button, 'Cancel')

When showing a message, the layout of controls are (vertical):
- Label (message text)
- Button Bar (Horizontal, 2 buttons, 'Try Again' and 'Cancel')

While implementing this I noticed that the info bubble in GTK did not redraw itself properly if the content increased in width or height after creation, which seemed to be because the new mask region was 'combined' with the previous region. So I've added a line to clear the previous mask region before setting a new one.

BUG=53598
TEST=manual, unplug mic and start recognition to check error message, and similarly give no speech to check.

Review URL: http://codereview.chromium.org/3307025

git-svn-id: http://src.chromium.org/svn/trunk/src@59433 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a51207568ec938b74d6d2734d38baa13c0567898