Add a scorer class for client-side phishing detection.
The Scorer class loads a machine learning model and lets you compute a
phishing score for a set of features that were extracted from the current
page.  The phishing score corresponds to the probability that the page is
phishing.  The features are extracted by the various feature extractor classes
that live in the same folder.  The current format of the model is a
protocol buffer (see: client_mode.proto).

BUG=NONE
TEST=PhishingScorerTest

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=59019

Review URL: http://codereview.chromium.org/3363004

git-svn-id: http://src.chromium.org/svn/trunk/src@59434 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 231ba0fd2b8d51d197f82ef7bfc3c50b9a5369f0