Mac: Fix concurrent access on a non-threadsafe data structure.
Also remove a call to cocoa from the non-ui thread.

DrawAcceleratedSurfaceInstance is called on the display link thread. plugin_views_ can be written to from the main thread. Since the map is only used to get the view's size, pass in the size right away.

There's another problem like this in plugin_container_manager_, which also has a map that is accessed from two threads. I will address this in a follow-up CL.

BUG=52491
TEST=none

Review URL: http://codereview.chromium.org/3431011

git-svn-id: http://src.chromium.org/svn/trunk/src@59684 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e680025e810d3089435c22b7e25ce09d1f8204da