Disk cache: Fix the order in which we delete data from the block files.
Stress testing the cache reveals a problem with the deletion
of some data from an entry: it is possible to crash in a way
that the block file thinks a block is free and an entry
thinks the block is in use. This CL corrects that issue.

There is also some new tests and a bunch of DCHECKS added
while looking for the problem, as well as adding tests to
make sure that a block file is accessed only from one thread
(there is no problem with the current code in that regard)

BUG=55605
TEST=netunittests

Review URL: http://codereview.chromium.org/3430004

git-svn-id: http://src.chromium.org/svn/trunk/src@59711 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 73cb05c741e871d73f26286d6a5647e236fdb77b