Disk cache: Fix the order in which we delete data from the block files.
Stress testing the cache reveals a problem with the deletion
of some data from an entry: it is possible to crash in a way
that the block file thinks a block is free and an entry
thinks the block is in use. This CL corrects that issue.

There is also some new tests and a bunch of DCHECKS added
while looking for the problem, as well as adding tests to
make sure that a block file is accessed only from one thread
(there is no problem with the current code in that regard)

BUG=55605
TEST=netunittests

Review URL: http://codereview.chromium.org/3430004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@59711 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a6c3aa9eab6a67a1cc32b878089d85ca926ee169