Support for building Chrome using Clang.
To build, set the clang=1 gyp_define.

This patch is the culmination of many months of effort and many
patches.  It contains the minimal changes to Chrome that
are Clang-specific.

With this, I can build the "chrome" target.  Once this patch
is in, we can incrementally fix bits of Chrome and various
tests and remove the Clang-specific workarounds.

Review URL: http://codereview.chromium.org/522020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@59882 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 67c193e5857d42852a66280ea3501635c4534264