DOMUI: Fix a bug in the AutoFill tab.
Bug: When there is no items in the profile list, click on the checkbox to
disable the autofill feature. Notice that all buttons become disable (that is ok).
Now click again to enable the feature, notice that all buttons become enabled
(that is NOT ok, because there aren't any items in the profile list).

Fix: Properly enable/disable the Edit and Remove buttons based on the profile
list size and the current state of the preference.

BUG=None
TEST=see Bug description above.

Review URL: http://codereview.chromium.org/3390016

git-svn-id: http://src.chromium.org/svn/trunk/src@59928 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7527e6f8ed7d3013421267c13d8b1bb13414f629