DOMUI: Fix a bug in the AutoFill tab.
Bug: When there is no items in the profile list, click on the checkbox to
disable the autofill feature. Notice that all buttons become disable (that is ok).
Now click again to enable the feature, notice that all buttons become enabled
(that is NOT ok, because there aren't any items in the profile list).

Fix: Properly enable/disable the Edit and Remove buttons based on the profile
list size and the current state of the preference.

BUG=None
TEST=see Bug description above.

Review URL: http://codereview.chromium.org/3390016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@59928 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f539ff87d993c8885ebc4b9a89bc0eba572845d8