Updating status on three tests:
Someone changed FindInPagePrematureEnd from DISABLED to FAILS. However, it is not expected to pass until we fix 11761 so we don't need to run it until then. Marking it back as DISABLED.

NoIncognitoPrepopulate was failing because the infrastructure was laggy (the simple act of launching a page in incognito could take up to a minute). I can no longer repro the error, so I think the problem is gone.

Back in June, ActivateLinkNavigatesPage got marked as DISABLED (because it crashed occasionally), with a bug filed against no-one, so I have no crashdump or try bot logs nor any idea what the state of it is. Marking as enabled to see if it is still a problem. It passes locally.

BUG=45594, 40948
TEST=These are tests.

Review URL: http://codereview.chromium.org/601044

git-svn-id: http://src.chromium.org/svn/trunk/src@59930 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1732aaa111568cb03da70261176b0d3fbaea006e