[Mac] Explicitly set Space when existing full screen
If a user moves the full screen window to a different space they get snapped back to the original space when they exit full screen. 

The problem is that internally the full screen window is a different window than the normal browser window. This means that any changes the user makes to the full screen window aren't reflected on the normal browser window. 

The correct fix for this bug is to create a new browser window when the user exits full screen. This new window would then be on the correct space. This would require a fairly large architectural change to the BrowserWindowController class. 

My proposed fix is much simpler. When exiting full screen we simply move the normal browser window to the Space that the full screen window is on. 

Original review at: http://codereview.chromium.org/3345005/show

Patch by Sailesh Agrawal <fnepal@gmail.com>.

BUG=41693
TEST=Moved the full screen window to a different Space. Verified that exiting full screen didn't change the active Space. 
Review URL: http://codereview.chromium.org/3386015

git-svn-id: http://src.chromium.org/svn/trunk/src@60063 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ff460c5575ac59968eeedd2f51f5f8d81d5009d2