Short-circuit safe-browsing updates once corruption is detected.
The corruption handler posts a task to the current message loop to
delete the database after the current task has completed.  The older
code re-created the database and allowed the transaction to continue,
the newer code did not, causing a crash.  [The database was deleted
first, so on next launch it would not crash again.]

This changes the new code to short-circuit storage of further chunks
once corruption is detected, until the next update (which will
re-create the database as needed).

Adding the tests exposed some code paths which weren't quite right.
Pulled the lock out of ClearUpdateCaches() because ResetDatabase()
calls it with a lock.  The corruption handler can delete open
databases.  Added sanity-check code to SafeBrowsingStoreFile so that
it doesn't allocate insane vectors.

BUG=55606
TEST=crash in 55606 stops happening, tests green.

Review URL: http://codereview.chromium.org/3420015

git-svn-id: http://src.chromium.org/svn/trunk/src@60128 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b1df0c6db93899a69edb3d5771186e4bbf6bd603