Cleanup orphaned testserver processes on posix.
If a chrome test that uses a testserver were to crash, the testserver process ends up being orphaned, potentially affecting subsequent tests. This patch causes child processes of a test case to be killed along with the test case if it were to be abruptly terminated.

BUG=55808
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/3423012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@60192 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8911c9a7978f04b879bd0adf95ffeb7a323b513c