Cleanup orphaned testserver processes on posix.
If a chrome test that uses a testserver were to crash, the testserver process ends up being orphaned, potentially affecting subsequent tests. This patch causes child processes of a test case to be killed along with the test case if it were to be abruptly terminated.

BUG=55808
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/3423012

git-svn-id: http://src.chromium.org/svn/trunk/src@60192 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 21e3137c7a784edde068e6f01ed1f5e94542d4cb