Improve flaky ui_test AutomationProxyVisibleTest.AutocompleteMatchesTest
The omnibox requires that it is focused before text input to it will
kick off the autocompletion providers. We add code
(edit->WaitForFocus()) to the test to wait for the focus to be obtained
before setting the omnibox’s text input and proceeding with the rest of
the test.

BUG=19876
TEST=Existing

Review URL: http://codereview.chromium.org/3348021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@60443 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: fa64e923f864cc023dd0b662741a81334d4a1695