Improve flaky ui_test AutomationProxyVisibleTest.AutocompleteMatchesTest
The omnibox requires that it is focused before text input to it will
kick off the autocompletion providers. We add code
(edit->WaitForFocus()) to the test to wait for the focus to be obtained
before setting the omnibox’s text input and proceeding with the rest of
the test.

BUG=19876
TEST=Existing

Review URL: http://codereview.chromium.org/3348021

git-svn-id: http://src.chromium.org/svn/trunk/src@60443 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 059ae691a6b562fdeea4b26dfde5f0b4237f3454