Clean up the policy and message model of the template generator
In the old model each policy had to be in a group, and messages could be inherited from groups to policies.

In the new model single policies does not have to be in groups and there is no inheritance of messages across policies/groups. Instead of that, policies have now two captions for different purposes. (Documented in policy_templates.json.) This is a more simple concept.

BUG=55722,56512
TEST=covered by policy template generator unit tests

Review URL: http://codereview.chromium.org/3454017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@60458 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d05a753f3e1f5f642058ed83150cc6284ec3af51