Revert 60556 (broke chrome, at least on linux) - Refactor, rename and relocate ProfileSyncServiceTestHarness.
ProfileSyncServiceTestHarness needs to support a specialization that can use a WaitableEvent instead of starting and stopping the UI Message loop. Required for automation provider support for sync.

The class also needs to live in chrome/browser, instead of chrome/test, and could do with a rename.

BUG=56460
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/3492005

TBR=rsimha@chromium.org

git-svn-id: http://src.chromium.org/svn/trunk/src@60584 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 61dff0d91c8e3893deefa932ea3090490ec250c1