Revert 60556 (broke chrome, at least on linux) - Refactor, rename and relocate ProfileSyncServiceTestHarness.
ProfileSyncServiceTestHarness needs to support a specialization that can use a WaitableEvent instead of starting and stopping the UI Message loop. Required for automation provider support for sync.

The class also needs to live in chrome/browser, instead of chrome/test, and could do with a rename.

BUG=56460
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/3492005

TBR=rsimha@chromium.org

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@60584 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 29c9c3afcfd6ae22a484f23ef1448be7df3b6649