Get the speech input browser test working again.
This test was having issues in Mac and was disabled. After that we also added a webkit change
to not allow programmatic events from Javascript to trigger speech input, so this test was not
working anymore.

In this CL I have rewritten the test to send the mouse click via the RenderView and this
should fix the flakiness and get it working again.

BUG=51337
TEST=SpeechInputBrowserTest.TestBasicRecognition on all platforms.

Review URL: http://codereview.chromium.org/3446029

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@60916 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 96bf77de9652e213519aa0290849e251a7598496