HttpAuthHandlerFactory::CreateDefault requires a HostResolver argument.
Previously the default registry factory would crash if it saw a Negotiate header when trying to do a CNAME lookup because of a NULL resolver. Since then I've changed HttpAuthHandlerNegotiate to skip the CNAME lookup step when the resolver_ is NULL.

So, one option is to make CreateDefault accept an optional HostResolver argument. My concern is that since passing in NULL is the easier choice, people will do it and encounter user problems when going through a Negotiate proxy. Hence, I force the correct behavior.

BUG=56819
TEST=net_unittests, try fetch client through a Negotiate authenticating proxy.

Review URL: http://codereview.chromium.org/3518001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@60942 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d4d5b38c53838ab3a4dc3c464fbebcd9fca8ce53