HttpAuthHandlerFactory::CreateDefault requires a HostResolver argument.
Previously the default registry factory would crash if it saw a Negotiate header when trying to do a CNAME lookup because of a NULL resolver. Since then I've changed HttpAuthHandlerNegotiate to skip the CNAME lookup step when the resolver_ is NULL.

So, one option is to make CreateDefault accept an optional HostResolver argument. My concern is that since passing in NULL is the easier choice, people will do it and encounter user problems when going through a Negotiate proxy. Hence, I force the correct behavior.

BUG=56819
TEST=net_unittests, try fetch client through a Negotiate authenticating proxy.

Review URL: http://codereview.chromium.org/3518001

git-svn-id: http://src.chromium.org/svn/trunk/src@60942 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4b3086499a773e80c6ec2e6ec28a534bc7f469fd