Revert 60949 - Render crash in FormManager::FindCachedFormElement()
To address the vulnerability of stale WebFrame pointers in the FormManager's cache this CL changes the cache from a map (with the WebFrame pointer as "key") to a flat vector of simplified "FormElement*" items.

To avoid leaking memory, we need to still observe |frameDetached|, and use that as a signal to reap any associated WebFormElements or WebFormControlElements.

BUG=48857
TEST=FormMananagerTest.*, and manual test of regular form filling, form filling a form with sub-iframes, and form filling a form with sub-frames.

Review URL: http://codereview.chromium.org/3492015

TBR=dhollowa@chromium.org
Review URL: http://codereview.chromium.org/3543003

git-svn-id: http://src.chromium.org/svn/trunk/src@60953 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4110a4929ff22a3912ada584f00b74b45c4d16f2