Revert r61095 and r61128. I found several issues when testing this and I don't feel comfortable leaving it on trunk while they are ironed out.
The terminology in the code is a bit schizophrenic at the
moment, so for clarity:

open = big icons showing
closed = small icons showing
hidden = no icons showing (button in bottom of tab)

1. If the apps section is hidden and an app is installed, the
apps section comes back weirdly. The content is visible, but
the header is not and the menu along the bottom of the tab
is still there.

2. Sections' open state not persisted. If you click on a
header to make a section open, then refresh, that section is
closed again.

3. If you open the most visited section, then refresh, the
most visited section will look closed. But if you hover in
the empty area in the bottom of the tab, you will get
cursor changes and status bubbles. This is probably related
to 2, above.

These aren't that hard to fix probably, but there are a few
other high priority minor fixes I need to get in for m7,
and I'm afraid of this change conflicting with them for
merges. So I'm pulling this out until one of us has time to
get back to it.

TBR=thakis@chromium.org
BUG=55148

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@61130 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 41b24fdd7b78d59088fc760350c64c3b838166ae