Fix autocomplete_unittest two-phase lookup problems.
The operator<< for AutocompleteResult::const_iterator needs to be declared
in the same namespace as its parameter type, in this case the global namespace.

Otherwise Clang fails to compile this because it is more strict on the Standard.

See http://clang.llvm.org/compatibility.html#dep_lookup for more info.

BUG=none
TEST=unit_tests

Review URL: http://codereview.chromium.org/3591001

git-svn-id: http://src.chromium.org/svn/trunk/src@61162 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 02d0e29ba340990072db07819cb104421d4457f9