Fix form_structure_unittest two-phase lookup problems.
The operator<< for webkit_glue::FormData needs to be declared
in the same namespace as its parameter type, in this case webkit_glue.

Otherwise Clang fails to compile this because it is more strict on the Standard.

See http://clang.llvm.org/compatibility.html#dep_lookup for more info.

BUG=none
TEST=unit_tests

Review URL: http://codereview.chromium.org/3583002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@61164 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: bd15ee5f11071bfa44a69b9189ba99a8339fa794