Fix url_fixer_upper_unittest two-phase lookup problems.
The operator<< for url_parse::Component needs to be declared
in the same namespace as its parameter type, in this case url_parse.

Otherwise Clang fails to compile this because it is more strict on the Standard.

See http://clang.llvm.org/compatibility.html#dep_lookup for more info.

BUG=none
TEST=unit_tests

Review URL: http://codereview.chromium.org/3575002

git-svn-id: http://src.chromium.org/svn/trunk/src@61167 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: cb872893633cecc51fa38a13b818cb7fc0d1c845